More bug fixes, display languages now display for the native speaker

+++ src/net/sf/memoranda/ui/EventsPanel.java
@@ -396,7 +396,7 @@ public class EventsPanel extends JPanel {
             ed = new CalendarDate((Date) dlg.endDate.getModel().getValue());
         if (dlg.dailyRepeatRB.isSelected()) {
             rtype = EventsManager.REPEAT_DAILY;
-            period = ((Integer) dlg.daySpin.getModel().getValue()).intValue();
+            period = ((Integer) dlg.daySpin.getModel().getValue());
         }
         else if (dlg.weeklyRepeatRB.isSelected()) {
             rtype = EventsManager.REPEAT_WEEKLY;
@@ -413,7 +413,7 @@ public class EventsPanel extends JPanel {
 	}
         else {
             rtype = EventsManager.REPEAT_MONTHLY;
-            period = ((Integer) dlg.dayOfMonthSpin.getModel().getValue()).intValue();
+            period = ((Integer) dlg.dayOfMonthSpin.getModel().getValue());
         }
         EventsManager.createRepeatableEvent(rtype, sd, ed, period, hh, mm, text, dlg.workingDaysOnlyCB.isSelected());
     }
